+++ src/commands/fun/other/suggest.js
@@ -3,7 +3,7 @@ const fs = require('fs');
 
 module.exports = {
     run: async (client, message, args) => {
-        message.delete();
+        if (message.guild.me.hasPermission('MANAGE_MESSAGES')) message.delete();
         let bannedWords = fs.readFileSync('./events/bannedwords.txt').toString().split("\r\n");
         let bannedPhrases = fs.readFileSync('./events/bannedphrases.txt').toString().split("\r\n");
         let msg = message.content.toLowerCase();
@@ -27,10 +27,9 @@ module.exports = {
             description: args,
             timestamp: new Date()
         }
-        message.delete();
         if (message.guild.id === '662734925707083778') {
             message.channel.send({ embed: suggestEmbed }).then(embedMessage => {
-                const yesEmoji = message.guild.emojis.cache.find(emoji => emoji.name === 'yep');
+                const yesEmoji = message.guild.emojis.cache.find(emoji => emoji.name === 'yes');
                 const noEmoji = message.guild.emojis.cache.find(emoji => emoji.name === 'nope');
                 embedMessage.react(yesEmoji);
                 embedMessage.react(noEmoji);
